# static/openapi.yaml
openapi: 3.0.0
info:
  title: La Maison API
  version: 1.0.0
servers:
  - url: http://127.0.0.1:4000
paths:
  /api/admin/login:
    post:
      tags:
        - default
      summary: 'ADMIN: Login admin'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                email: ''
                password: ''
      parameters:
        - name: Content-Type
          in: header
          schema:
            type: string
          example: application/json
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...
  /api/login:
    post:
      tags:
        - default
      summary: 'USER: Login user'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                email: ''
                password: ''
      parameters:
        - name: Content-Type
          in: header
          schema:
            type: string
          example: application/json
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...